requirements.txt
psycopg2-binary 
python-dotenv

ollama
pdfplumber 
langchain
langchain-core
langchain-ollama
langchain-community
langchain_text_splitters

pandas
